requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
opencv-python-headless
Pillow
